+++ fe/__tests__/index/filter.test.js
@@ -1,0 +1,16 @@
+describe('when using filter functions', () => {
+  it.todo('should return correct rows in the table')
+  it.todo('should fire search event after pressing enter key')
+  it.todo('should fire search event when field is cleared')
+  it.todo('should not fire search event twice or more if value is same')
+  it.todo('should jump back to page 1 if search is triggered')
+  it.todo('should retain filter values if search event is fired')
+  it.todo('cursor should retain filter values after firing search event')
+
+  it.todo('should fire filter event if option is changed')
+  it.todo('should not fire filter event twice or more if value is same')
+  it.todo('should jump back to page 1 if filter is change')
+  it.todo('should retain search values if filter event is fired')
+
+  it.todo('should clear search and filter values if reset event is fired')
+})
